Complete MBv2 quantization

+++ src/layer/quantize.h
@@ -30,8 +30,8 @@ public:
     virtual int forward(const Mat &bottom_blob, Mat &top_blob, const Option &opt) const;
 
 public:
-    // float scale;
-    int scale;
+    float scale;
+    //int scale;
     int position_bottom_scale;
     int position_scale_in;
 };
